requirements.txt
pandas
seaborn
peekingduck
pytorch
torchvision
cudatoolkit=11.3